fixed users tg_id bug #2

+++ core/models/Users.py
@@ -5,7 +5,7 @@ from .Favorites import favorites_table
 
 
 class User(Base):
-    tg_id: Mapped[BigInteger] = mapped_column(Integer, unique=True, index=True, nullable=False)
+    tg_id: Mapped[BigInteger] = mapped_column(BigInteger, unique=True, index=True, nullable=False)
     username: Mapped[str]
     balance = mapped_column(Integer, default=0)
     favorites: Mapped[list["Product"]] = relationship(
